requirements.txt
# requirements.txt
discord.py
dotenv
youtube_dl 
discord.py
dotenv
youtube_dl